Fixed packet type of version packet
Updated the DIFI dissector to correctly interpret packet types and classes as per the official
DIFI specification. We now accurately recognize that the protocol defines **only three top-level
packet types: Data (0x1), Context (0x4), and Command (0x6)**. The "Version" packet is
specifically identified as **Class 4 within the Context (0x4) packet type**.
The value of 5 for the packet type was only a typo in the table of the version packet.